Revert "readDMIData + Read BIOS Vendor check"
